Break more controls out to files
