+++ Feedback.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+type FeedbackType = {
+  showFeedback : (positive: boolean) => void;
+}
+
+const FeedbackContext = React.createContext<FeedbackType>({showFeedback: () => {}});
+
+export { FeedbackContext }
