+++ Settings.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+type SettingsType = {
+  apiKey?: string;
+}
+  
+const SettingsContext = React.createContext<SettingsType>({});
+
+export { SettingsContext }
